adding mongoose find
